backend/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
